+++ GloryEngine/Editor/GloryEditor/AssetReferencePropertyDrawer.h
@@ -10,9 +10,9 @@ namespace Glory::Editor
 		AssetReferencePropertyDrawer() : PropertyDrawer(SerializedType::ST_Asset) {}
 
 	protected:
-		virtual bool Draw(const SerializedProperty* serializedProperty, const std::string& label, void* data, size_t typeHash, uint32_t flags) const override;
+		virtual bool Draw(const SerializedProperty* serializedProperty, const std::string& label, void* data, uint32_t typeHash, uint32_t flags) const override;
 		virtual bool Draw(const SerializedProperty* serializedProperty, const std::string& label, std::any& data, uint32_t flags) const override;
-		virtual bool Draw(const std::string& label, void* data, size_t typeHash, uint32_t flags) const override;
-		virtual bool Draw(const std::string& label, YAML::Node& node, size_t typeHash, uint32_t flags) const;
+		virtual bool Draw(const std::string& label, void* data, uint32_t typeHash, uint32_t flags) const override;
+		virtual bool Draw(const std::string& label, YAML::Node& node, uint32_t typeHash, uint32_t flags) const;
 	};
 }
