+++ GloryEngine/Engine/GloryCore/MaterialPropertyInfo.cpp
@@ -2,13 +2,13 @@
 
 namespace Glory
 {
-	MaterialPropertyInfo::MaterialPropertyInfo(const std::string& displayName, const std::string& shaderName, size_t typeHash, size_t size, size_t offset, uint32_t flags)
+	MaterialPropertyInfo::MaterialPropertyInfo(const std::string& displayName, const std::string& shaderName, uint32_t typeHash, size_t size, size_t offset, uint32_t flags)
 		: m_PropertyDisplayName(displayName), m_PropertyShaderName(shaderName), m_TypeHash(typeHash),
 		m_Size(size), m_Offset(offset), m_IsResource(false), m_Flags(flags)
 	{
 	}
 
-	MaterialPropertyInfo::MaterialPropertyInfo(const std::string& displayName, const std::string& shaderName, size_t typeHash, size_t offset, uint32_t flags)
+	MaterialPropertyInfo::MaterialPropertyInfo(const std::string& displayName, const std::string& shaderName, uint32_t typeHash, size_t offset, uint32_t flags)
 		: m_PropertyDisplayName(displayName), m_PropertyShaderName(shaderName), m_TypeHash(typeHash),
 		m_Size(sizeof(UUID)), m_Offset(offset), m_IsResource(true), m_Flags(flags)
 	{
@@ -24,7 +24,7 @@ namespace Glory
 		return m_PropertyShaderName;
 	}
 
-	const size_t MaterialPropertyInfo::TypeHash() const
+	const uint32_t MaterialPropertyInfo::TypeHash() const
 	{
 		return m_TypeHash;
 	}
